Remove local.py, rename exaples accordingly
