+++ examples/runner-examples/02-two-transport-bot.py
@@ -6,14 +6,7 @@
 
 """Pipecat Cloud-compatible bot example.
 
-Transports are:
-
-- Daily
-- SmallWebRTC
-- Twilio
-- Telnyx
-- Plivo
-"""
+Transports are Daily or SmallWebRTC."""
 
 import os
 
@@ -26,20 +19,21 @@ from pipecat.pipeline.runner import PipelineRunner
 from pipecat.pipeline.task import PipelineParams, PipelineTask
 from pipecat.processors.aggregators.openai_llm_context import OpenAILLMContext
 from pipecat.processors.frameworks.rtvi import RTVIConfig, RTVIObserver, RTVIProcessor
-from pipecat.runner.cloud import SmallWebRTCSessionArguments
+from pipecat.runner.run import SmallWebRTCSessionArguments
 from pipecat.services.cartesia.tts import CartesiaTTSService
 from pipecat.services.deepgram.stt import DeepgramSTTService
 from pipecat.services.openai.llm import OpenAILLMService
 
 try:
-    from pipecatcloud.agent import DailySessionArguments, WebSocketSessionArguments
+    from pipecatcloud.agent import DailySessionArguments
 except ImportError:
     raise ImportError(
         "pipecatcloud package is required for cloud-compatible bots. "
-        "Install with: pip install pipecat-ai[pipecatcloud]"
+        "Install with: pip install pipecat-ai[[pipecatcloud]]"
     )
 
 load_dotenv(override=True)
+
 
 # Check if we're running locally
 IS_LOCAL_RUN = os.environ.get("LOCAL_RUN", "0") == "1"
@@ -107,9 +101,7 @@ async def run_bot(transport):
     await runner.run(task)
 
 
-async def bot(
-    session_args: DailySessionArguments | SmallWebRTCSessionArguments | WebSocketSessionArguments,
-):
+async def bot(session_args: DailySessionArguments | SmallWebRTCSessionArguments):
     """Main bot entry point compatible with Pipecat Cloud."""
 
     if isinstance(session_args, DailySessionArguments):
@@ -145,67 +137,10 @@ async def bot(
             webrtc_connection=session_args.webrtc_connection,
         )
 
-    elif isinstance(session_args, WebSocketSessionArguments):
-        # Use the utility to parse WebSocket data
-        from pipecat.runner.utils import parse_telephony_websocket
-
-        transport_type, stream_id, call_id = await parse_telephony_websocket(session_args.websocket)
-        logger.info(f"Auto-detected transport: {transport_type}")
-
-        # Create transport based on detected type
-        if transport_type == "twilio":
-            from pipecat.serializers.twilio import TwilioFrameSerializer
-
-            serializer = TwilioFrameSerializer(
-                stream_sid=stream_id,
-                call_sid=call_id,
-                account_sid=os.getenv("TWILIO_ACCOUNT_SID", ""),
-                auth_token=os.getenv("TWILIO_AUTH_TOKEN", ""),
-            )
-
-        elif transport_type == "telnyx":
-            from pipecat.serializers.telnyx import TelnyxFrameSerializer
-
-            serializer = TelnyxFrameSerializer(
-                stream_id=stream_id,
-                call_control_id=call_id,
-                outbound_encoding="PCMU",  # Set manually
-                inbound_encoding="PCMU",
-            )
-
-        elif transport_type == "plivo":
-            from pipecat.serializers.plivo import PlivoFrameSerializer
-
-            serializer = PlivoFrameSerializer(
-                stream_id=stream_id,
-                call_id=call_id,
-            )
-
-        else:
-            # Generic fallback
-            serializer = None
-
-        # Create the transport
-        from pipecat.transports.network.fastapi_websocket import (
-            FastAPIWebsocketParams,
-            FastAPIWebsocketTransport,
-        )
-
-        transport = FastAPIWebsocketTransport(
-            websocket=session_args.websocket,
-            params=FastAPIWebsocketParams(
-                audio_in_enabled=True,
-                audio_out_enabled=True,
-                add_wav_header=False,
-                vad_analyzer=SileroVADAnalyzer(),
-                serializer=serializer,
-            ),
-        )
-
     await run_bot(transport)
 
 
 if __name__ == "__main__":
-    from pipecat.runner.cloud import main
+    from pipecat.runner.run import main
 
     main()
