+++ examples/runner-examples/Dockerfile
@@ -4,4 +4,4 @@ COPY ./requirements.txt requirements.txt
 
 RUN pip install --no-cache-dir --upgrade -r requirements.txt
 
-COPY ./cloud-simple-bot.py bot.py
+COPY ./02-two-transport-bot.py bot.py
