--- examples/runner-examples/local-multi-bot.py
@@ -1,122 +1,0 @@
-#
-# Copyright (c) 2024–2025, Daily
-#
-# SPDX-License-Identifier: BSD 2-Clause License
-#
-
-import argparse
-import os
-
-from dotenv import load_dotenv
-from loguru import logger
-
-from pipecat.audio.vad.silero import SileroVADAnalyzer
-from pipecat.pipeline.pipeline import Pipeline
-from pipecat.pipeline.runner import PipelineRunner
-from pipecat.pipeline.task import PipelineParams, PipelineTask
-from pipecat.processors.aggregators.openai_llm_context import OpenAILLMContext
-from pipecat.processors.frameworks.rtvi import RTVIConfig, RTVIObserver, RTVIProcessor
-from pipecat.services.cartesia.tts import CartesiaTTSService
-from pipecat.services.deepgram.stt import DeepgramSTTService
-from pipecat.services.openai.llm import OpenAILLMService
-from pipecat.transports.base_transport import BaseTransport, TransportParams
-
-load_dotenv(override=True)
-
-
-def create_transport_params(transport_name):
-    """Create transport parameters based on transport name."""
-    base_config = {
-        "audio_in_enabled": True,
-        "audio_out_enabled": True,
-        "vad_analyzer": SileroVADAnalyzer(),
-    }
-
-    if transport_name == "daily":
-        from pipecat.transports.services.daily import DailyParams
-
-        return DailyParams(**base_config)
-    elif transport_name == "livekit":
-        from pipecat.transports.services.livekit import LiveKitParams
-
-        return LiveKitParams(**base_config)
-    elif transport_name in ["plivo", "telnyx", "twilio"]:
-        from pipecat.transports.network.fastapi_websocket import FastAPIWebsocketParams
-
-        return FastAPIWebsocketParams(**base_config)
-    else:  # webrtc
-        return TransportParams(**base_config)
-
-
-async def run_bot(transport: BaseTransport, _: argparse.Namespace, handle_sigint: bool):
-    logger.info(f"Starting bot")
-
-    stt = DeepgramSTTService(api_key=os.getenv("DEEPGRAM_API_KEY"))
-
-    tts = CartesiaTTSService(
-        api_key=os.getenv("CARTESIA_API_KEY"),
-        voice_id="71a7ad14-091c-4e8e-a314-022ece01c121",  # British Reading Lady
-    )
-
-    llm = OpenAILLMService(api_key=os.getenv("OPENAI_API_KEY"))
-
-    messages = [
-        {
-            "role": "system",
-            "content": "You are a friendly AI assistant. Respond naturally and keep your answers conversational.",
-        },
-    ]
-
-    context = OpenAILLMContext(messages)
-    context_aggregator = llm.create_context_aggregator(context)
-
-    rtvi = RTVIProcessor(config=RTVIConfig(config=[]))
-
-    pipeline = Pipeline(
-        [
-            transport.input(),  # Transport user input
-            rtvi,
-            stt,
-            context_aggregator.user(),  # User responses
-            llm,  # LLM
-            tts,  # TTS
-            transport.output(),  # Transport bot output
-            context_aggregator.assistant(),  # Assistant spoken responses
-        ]
-    )
-
-    task = PipelineTask(
-        pipeline,
-        params=PipelineParams(
-            enable_metrics=True,
-            enable_usage_metrics=True,
-        ),
-        observers=[RTVIObserver(rtvi)],
-    )
-
-    @transport.event_handler("on_client_connected")
-    async def on_client_connected(transport, client):
-        logger.info(f"Client connected")
-        # Kick off the conversation.
-        messages.append({"role": "system", "content": "Say hello and briefly introduce yourself."})
-        await task.queue_frames([context_aggregator.user().get_context_frame()])
-
-    @transport.event_handler("on_client_disconnected")
-    async def on_client_disconnected(transport, client):
-        logger.info(f"Client disconnected")
-        await task.cancel()
-
-    runner = PipelineRunner(handle_sigint=handle_sigint)
-
-    await runner.run(task)
-
-
-if __name__ == "__main__":
-    from pipecat.runner.local import main
-
-    transport_params = {
-        transport: lambda t=transport: create_transport_params(t)
-        for transport in ["daily", "livekit", "plivo", "telnyx", "twilio", "webrtc"]
-    }
-
-    main(run_bot, transport_params=transport_params)
